Move on enter to a separate module
